#390 refactor(move to integrationTest)
