+++ build.gradle
@@ -189,36 +189,31 @@ webdriverBinaries {
     geckodriver "$geckoDriverVersion"
 }
 
-drivers.each { driver ->
-    task "${driver}Test"(type: Test) {
-        group JavaBasePlugin.VERIFICATION_GROUP
 
-        outputs.upToDateWhen { false }  // Always run tests
+tasks.getByName("integrationTest") {
+    group JavaBasePlugin.VERIFICATION_GROUP
 
-        systemProperty "geb.build.reportsDir", reporting.file("geb/$name")
-        systemProperty "geb.env", driver
+    //outputs.upToDateWhen { false }  // Always run tests
 
-        systemProperty "geb.build.baseUrl", System.properties['baseUrl'] ?: "http://devt.ala.org.au:8087"  // pass test server by -DbaseUrl
+    systemProperty "geb.build.reportsDir", reporting.file("geb/$name")
 
-        //Store username and password into the default config file
-        systemProperty "configFile",System.properties['configFile']?: "/data/spatial-hub/test/default.properties"
-
-        systemProperty "username", System.properties['username']
-        systemProperty "password", System.properties['password']
-
-    }
+    systemProperty "geb.env",  System.properties['driver']?:"firefox"  // firefox by default
+    // pass test server by -DbaseUrl
+    systemProperty "geb.build.baseUrl", System.properties['baseUrl'] ?: "http://devt.ala.org.au:8087"
+    //Store username and password into the default config file
+    systemProperty "configFile",System.properties['configFile']?: "/data/spatial-hub/test/default.properties"
+    //or pass with -Dusername, -Dpassword
+    systemProperty "username", System.properties['username']
+    systemProperty "password", System.properties['password']
 }
 
-test {
-    dependsOn drivers.collect { tasks["${it}Test"] }
-    enabled = false
-}
 
 tasks.withType(Test) {
     maxHeapSize = "1g"
     jvmArgs '-XX:MaxMetaspaceSize=128m'
     testLogging {
         exceptionFormat = 'full'
+        showStandardStreams = true
     }
 }
