+++ src/integration-test/groovy/AddFacetSpec.groovy
