+++ src/integration-test/groovy/AddLayerSpec.groovy
