+++ src/integration-test/groovy/AddSpeciesSpec.groovy
