+++ src/integration-test/groovy/ToolClassifySpec.groovy
