+++ src/integration-test/groovy/ToolCompareAreaSpec.groovy
