+++ src/integration-test/groovy/ToolComparePointsSpec.groovy
