+++ src/integration-test/groovy/ToolDoubleGridingSpec.groovy
