+++ src/integration-test/groovy/ToolGeneratePointsSpec.groovy
