+++ src/integration-test/groovy/ToolNearestLocalitySpec.groovy
