+++ src/integration-test/groovy/ToolPointsToGridSpec.groovy
