+++ src/integration-test/groovy/ToolPredictSpec.groovy
