+++ src/integration-test/groovy/ToolReportSpec.groovy
