+++ src/integration-test/groovy/ToolScatterPlotSpec.groovy
