+++ src/integration-test/groovy/page/AddAreaModule.groovy
