+++ src/integration-test/groovy/page/AddFacetModule.groovy
