+++ src/integration-test/groovy/page/LegendModule.groovy
