+++ src/integration-test/groovy/page/MapModule.groovy
