+++ src/integration-test/groovy/page/SpatialHubHomePage.groovy
