+++ src/integration-test/groovy/page/ToolComparePointsModule.groovy
