+++ src/integration-test/resources/Functional_Test_Module.png
