+++ src/integration-test/resources/koppen.kml
